refactor(app.js): rename App.js to app.js
