+++ src/containers/root.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Provider } from 'react-redux'
-import App from './App'
+
+import App from './app'
 
 const Root = ({ store }) => {
   return <Provider store={store}>
